rename Fungible Token Standard to Contract
