+++ src/FungibleTokenContract.ts
