+++ src/test/approve.test.ts
@@ -17,7 +17,7 @@ import {
   FungibleToken,
   FungibleTokenErrors,
   VKeyMerkleMap,
-} from '../FungibleTokenStandard.js';
+} from '../FungibleTokenContract.js';
 import {
   MintConfig,
   MintParams,
@@ -241,7 +241,7 @@ describe('New Token Standard ApproveBase Tests', () => {
 
         await tokenContract.deploy({
           symbol: 'tokA',
-          src: 'https://github.com/o1-labs-XT/fungible-token-standard',
+          src: 'https://github.com/o1-labs-XT/fungible-token-contract',
         });
       });
